Moved the CPU module to a separate thread. Now emulator disassembles only ROM, not whole memory.

+++ src/disasm.h
@@ -19,7 +19,7 @@ class DisasmWidget : public QTextEdit
     Memory *memory;
 
 public slots:
-    void highlightCurrentLine(uint16_t pc);
+    void highlightCurrentLine(int pc);
     void cleanHighlight();
 
 public:
